feat(ios): share/import voice cards; mosaic calibration column and whitening
Lab:
- Share action on enrolled tiles opens VoiceCardSheet (preview + PNG share
  with plain-language explanation that the picture is the voice)
- PhotosPicker import decodes VoicePrintCard (chunk then mosaic); dedupes
  identical vectors into existing library entries
- Consent/library copy covers picture import; DEBUG FTTS_DEBUG_CARD opens
  a synthetic card for simulator screenshots

VoiceCode:
- Vertical calibration column (column 8) pins vertical scale the way the
  calibration row pins horizontal
- Whitening of the coded stream so float byte patterns do not stripe the
  field or starve threshold estimation
- Rank multiple finder triples best-first (data cells can imitate finders);
  CRC/calibration reject impostors
- Sub-pixel finder refinement, bilinear sampling, Float luma buffer

VoicePrintCard: CGFloat size cast for ImageRenderer proposedSize.

+++ ios/Sources/VoicePrintCard.swift
@@ -28,7 +28,8 @@ enum VoicePrintCard {
         let mosaic = mosaicImage(name: name, vector: vector)
         let renderer = ImageRenderer(content: CardView(name: name, mosaic: mosaic))
         renderer.scale = 1
-        renderer.proposedSize = .init(width: cardWidth, height: cardHeight)
+        renderer.proposedSize = .init(
+            width: CGFloat(cardWidth), height: CGFloat(cardHeight))
         guard let image = renderer.uiImage, let png = image.pngData() else {
             throw EngineError.native("cannot render the voice card")
         }
